Create getEmployerData function
