+++ src/pages/home/profile.tsx
@@ -1,9 +1,4 @@
 import { Layout } from "~/component/layout/layout";
-import { appRouter } from "~/server/api/root";
-import { prisma } from "~/server/db";
-import { AUTHORIZATION_TOKEN_KEY } from "~/utils/auth/authorizationTokenKey";
-import { verifyToken } from "~/utils/auth/auth";
-import type { VerifyToken } from "~/utils/auth/withoutAuth";
 import type {
   GetServerSidePropsContext,
   InferGetServerSidePropsType,
@@ -15,6 +10,7 @@ import { FormInput } from "~/component/profileForm/formInput";
 import type { CandidateFields as UserFields } from "~/component/candidate/candidateAccountForm";
 import { EmployerNavigationMenu } from "~/component/employer/employerNavigationMenu";
 import Head from "next/head";
+import { getEmployerData } from "~/utils/getEmployerData/getEmployerData";
 
 type EmployerData =
   | User & {
@@ -142,33 +138,8 @@ export default function EmployerProfile({
   );
 }
 
-export const getServerSideProps = async ({
-  req,
-}: GetServerSidePropsContext) => {
-  const authToken = req?.cookies[AUTHORIZATION_TOKEN_KEY] ?? "";
-  const verifiedToken = verifyToken(authToken) as VerifyToken | null;
-
-  const caller = appRouter.createCaller({ prisma });
-  const employerData = await caller.employer.findEmployeeById({
-    employerId: verifiedToken?.userId ?? "",
-  });
-
-  const isUserEmployer = employerData?.userType === "EMPLOYER";
-
-  if (!verifiedToken || !isUserEmployer) {
-    return {
-      redirect: {
-        destination: "/",
-        permanent: false,
-      },
-    };
-  }
-
-  const serializedEmployerData = superjson.stringify(employerData);
-
-  return {
-    props: {
-      employer: serializedEmployerData,
-    },
-  };
+export const getServerSideProps = async (
+  context: GetServerSidePropsContext
+) => {
+  return await getEmployerData(context);
 };
